% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{clear_cppad_mem}
\alias{clear_cppad_mem}
\title{Clears the Memory used by CppAD
Clear all pointers to CppAD objects and deallocate all memory used by
CppAD. This will make all subsequent use C++ pointer invalid.}
\usage{
clear_cppad_mem(max_n_threads = 1L, keep_work_space = FALSE)
}
\arguments{
\item{max_n_threads}{maximum number of threads which have been used.}

\item{keep_work_space}{logical for whether to keep}
}
\value{
An integer which is one if all memory could be freed.
}
\description{
Clears the Memory used by CppAD
Clear all pointers to CppAD objects and deallocate all memory used by
CppAD. This will make all subsequent use C++ pointer invalid.
}
