% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gsm_ADFun.R
\name{cov_to_theta}
\alias{cov_to_theta}
\alias{theta_to_cov}
\title{Maps Between a Covariance Matrix and Its Log-Cholesky Parametrization}
\usage{
cov_to_theta(cov)

theta_to_cov(theta)
}
\arguments{
\item{cov}{the covariance matrix.}

\item{theta}{numeric vector where the
first \eqn{K} elements are the log of \eqn{psi} and the
remaining \eqn{K (K - 1) / 2} elements are the non-zero and
non-diagonal entries of \eqn{H}.}
}
\description{
Let \eqn{C} be a \eqn{K}-dimensional covariance matrix. Let
\eqn{C = LL^\top} where \eqn{L} is the Cholesky decomposition
and \eqn{L = diag(\psi) H} where  \eqn{H} is a lower triangular matrix
with ones in the diagonal. Then the two functions map between \eqn{C} and
a vector where the first \eqn{K} elements are the log of \eqn{psi} and
the remaining \eqn{K (K - 1) / 2} elements are the non-zero and
non-diagonal entries of \eqn{H}.
}
\examples{
set.seed(1)
Sigma <- drop(rWishart(1, 4, diag(4)))
log_chol <- cov_to_theta(Sigma)
log_chol
stopifnot(all.equal(Sigma, theta_to_cov(log_chol)))

}
