% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gsm_ADFun.R
\name{cp_to_dp}
\alias{cp_to_dp}
\alias{dp_to_cp}
\title{Maps Between Centralized and Direct Parameters}
\usage{
cp_to_dp(mu, Sigma, gamma)

dp_to_cp(xi, Psi, alpha)
}
\arguments{
\item{mu}{mean of the random variable.}

\item{Sigma}{covariance matrix of the random variable.}

\item{gamma}{pearson skewness coefficients of the random variable.}

\item{xi}{location parameter of the random variable.}

\item{Psi}{scale parameter of the random variable.}

\item{alpha}{skewness parameter of the random variable.}
}
\description{
Maps between centralized parameters (CP) and direct parameters (DP) for
the multivariate skew-normal distribution.
}
\details{
We say that a random variable is skew-normal distributed if its density
is given by

\deqn{2\phi^{(K)}(y - \xi; \Psi)\Phi(\alpha^\top diag(\psi)^{-1}(y - \xi))}

where \eqn{\psi} is the square root of the diagonal of \eqn{\Psi}. The
two functions maps between the above DP to the CP where
\eqn{\mu = E(Y)}, \eqn{\Sigma = Var(Y)}, and
\eqn{\gamma_i =  E((Y_i - E(Y_i))^3) / Var(Y_i)^{3/2}}.
}
\examples{
mu <- 1:3
Sig <- diag(3)
Sig[lower.tri(Sig)] <- Sig[upper.tri(Sig)] <- .25
gamma <- ((1:3) - 2) / 10

dps <- cp_to_dp(mu = mu, Sigma = Sig, gamma = gamma)
dps
cps <- dp_to_cp(xi = dps$xi, Psi = dps$Psi, alpha = dps$alpha)

stopifnot(all.equal(mu   , cps$mu))
stopifnot(all.equal(Sig  , cps$Sigma))
stopifnot(all.equal(gamma, cps$gamma))

}
