% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mgsm_ADFun.R
\name{free_laplace}
\alias{free_laplace}
\title{Free Memory Used by TMB}
\usage{
free_laplace(obj)
}
\description{
Essentially a wrapper around \code{\link{FreeADFun}}.

 @param obj object of type MGSM_ADFun.
}
\seealso{
\code{\link{make_mgsm_ADFun}}
}
