% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint.R
\name{make_joint_ADFun}
\alias{make_joint_ADFun}
\title{Construct Objective Functions with Derivatives for a Joint Survival and
Marker Model}
\usage{
make_joint_ADFun(
  sformula,
  mformula,
  sdata,
  mdata,
  id_var,
  time_var,
  mknots,
  sknots,
  n_nodes = 20L,
  B = NULL,
  Psi = NULL,
  Sigma = NULL,
  omega = NULL,
  alpha = NULL,
  delta = NULL
)
}
\arguments{
\item{sformula}{\code{\link{formula}} for the time to event. The
left-hand side must be \code{\link{Surv}} object with
type "counting".}

\item{mformula}{\code{\link{formula}} for the marker process. The
left-hand side must \code{\link{cbind}}ed if there are
more than one marker.}

\item{sdata}{\code{\link{data.frame}} with the time to event data.}

\item{mdata}{\code{\link{data.frame}} with the marker data.}

\item{id_var}{name of the individual identifier. Must be present in both
\code{sdata} and \code{mdata}.}

\item{time_var}{name of the time variable in \code{mdata}.}

\item{mknots}{vector with knots for the baseline hazard or number of
knots to use.}

\item{n_nodes}{number of nodes to use with Gauss-Legendre quadrature for
the cumulative hazard.}

\item{B}{staring value for B.}

\item{Psi}{staring value for Psi.}

\item{Sigma}{staring value for Sigma.}

\item{omega}{staring value for omega.}

\item{alpha}{staring value for alpha.}

\item{delta}{staring value for delta.}
}
\description{
Construct Objective Functions with Derivatives for a Joint Survival and
Marker Model
}
