% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gsm_ADFun.R
\name{make_mgsm_ADFun}
\alias{make_mgsm_ADFun}
\title{Construct Objective Functions with Derivatives for a Mixed Generalized
Survival Model}
\usage{
make_mgsm_ADFun(
  formula,
  data,
  df,
  Z,
  cluster,
  do_setup = c("Laplace", "GVA", "SNVA"),
  n_nodes = 20L,
  param_type = c("DP", "CP_trans", "CP"),
  link = c("PH", "PO", "probit"),
  theta = NULL,
  beta = NULL,
  opt_func = .opt_default,
  n_threads = 1L,
  skew_start = -0.02,
  dense_hess = FALSE,
  sparse_hess = FALSE
)
}
\arguments{
\item{formula}{two-sided \code{\link{formula}} where the left-hand side is a
\code{\link{Surv}} object and the right-hand side is the
fixed effects.}

\item{data}{\code{\link{data.frame}} with variables used in the model.}

\item{df}{integer with the degrees of freedom used for the baseline
spline.}

\item{Z}{one-sided \code{\link{formula}} where the right-hand side are
the random effects.}

\item{cluster}{vector with integers or factors for group identifiers
(one for each observation).}

\item{do_setup}{character vector indicating which approximation to setup.
See 'Details'.}

\item{n_nodes}{integer with the number of nodes to use in (adaptive)
Gauss-Hermite quadrature.}

\item{param_type}{characters for the parameterization used with the SNVA.
See 'Details'.}

\item{link}{character specifying the link function.}

\item{theta}{starting values for covariance matrix.}

\item{beta}{starting values for fixed effect coefficients.}

\item{opt_func}{general optimization function to use. It
needs to have an interface like \code{\link{optim}}.}

\item{n_threads}{integer with number of threads to use.}

\item{skew_start}{starting value for the Pearson's moment coefficient of
skewness parameter when a SNVA is used. Currently,
a somewhat arbitrary value.}

\item{dense_hess}{logical for whether to make dense Hessian computation
available. Memory and computation time is saved if it is
\code{FALSE}.}

\item{sparse_hess}{logical for whether to make sparse Hessian computation
available. Memory and computation time is saved if it is
\code{FALSE}.}
}
\value{
An object of class \code{MGSM_ADFun}. The elements are:
\item{laplace}{object to perform a Laplace approximation if \code{do_setup} contains \code{"Laplace"}.}
\item{gva}{object to perform a GVA if \code{do_setup} contains \code{"GVA"} or \code{"SNVA"}.}
\item{snva}{object to perform a SNVA if \code{do_setup} contains \code{"SNVA"}.}
\item{y}{numeric vector with outcomes.}
\item{event}{numeric vector with event indicators.}
\item{X}{fixed effect design matrix.}
\item{XD}{derivative of fixed effect design matrix with respect to time.}
\item{Z}{Random effect design matrix.}
\item{grp}{integer vector with group identifier.}
\item{terms}{\code{\link{list}} with \code{\link{terms.object}}s.}
\item{link}{character with the link function.}
\item{cl}{matched call.}
\item{dense_hess}{dense_hess argument.}
\item{sparse_hess}{sparse_hess argument.}
}
\description{
Constructs an objective function for a particular generalized survival
model applied to a given data set.
}
\details{
Possible link functions for \code{link} are:
\code{"PH"} for proportional hazard,
\code{"PO"} for proportional odds, and \code{"probit"} for a probit link
function.

The available estimation methods for \code{do_setup} are:
\code{"Laplace"} for a Laplace approximation,
\code{"GVA"} for a Gaussian variational approximation (GVA),
and \code{"SNVA"} for a skew-normal variational approximation (SNVA).

The parameterizations for the SNVA are selected by \code{param_type}.
Possible arguments are: \code{"DP"} for direct parameterization,
\code{"CP"} for centralized parameterization, and
\code{"CP_trans"} for transformed centralized parameterization where the
skew parameters are transformed by a logistic function to be in the
appropriate range.

See the README \url{https://github.com/boennecd/survTMB} for
further information and examples.
}
\examples{
library(survTMB)
if(require(coxme)){
  # construct function with a random intercept and a proportional hazard
  # link function
  func <- make_mgsm_ADFun(
    Surv(y, uncens) ~ trt, cluster = as.factor(center), Z = ~ 1,
    df = 3L, data = eortc, link = "PH",
    do_setup = c("Laplace", "GVA", "SNVA"), n_threads = 1L)
  print(func)
}

}
\seealso{
\code{\link{fit_mgsm}}
}
