% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgsm-psqn.R
\name{make_mgsm_psqn_obj}
\alias{make_mgsm_psqn_obj}
\title{Create mgsm Object to Pass to psqn Method}
\usage{
make_mgsm_psqn_obj(
  formula,
  data,
  df,
  tformula = NULL,
  Z,
  cluster,
  method = c("SNVA", "GVA"),
  n_nodes = 20L,
  link = c("PH", "PO", "probit"),
  theta = NULL,
  beta = NULL,
  opt_func = .opt_default,
  n_threads = 1L,
  skew_start = -1e-04,
  kappa = .MGSM_default_kappa,
  dtformula = NULL
)
}
\arguments{
\item{formula}{two-sided \code{\link{formula}} where the left-hand side is a
\code{\link{Surv}} object and the right-hand side is the
fixed effects.}

\item{data}{\code{\link{data.frame}} with variables used in the model.}

\item{df}{integer with the degrees of freedom used for the baseline
spline.}

\item{tformula}{\code{\link{formula}} with baseline survival function.
The time variable must be the same
symbol as used in the left-hand-side of \code{formula}.
\code{NULL} implies that \code{df} is passed to
\code{\link{nsx}}.}

\item{Z}{one-sided \code{\link{formula}} where the right-hand side are
the random effects.}

\item{cluster}{vector with integers or factors for group identifiers
(one for each observation).}

\item{method}{Method character vector indicating which approximation to setup.
See \code{\link{make_mgsm_ADFun}}.}

\item{n_nodes}{integer with the number of nodes to use in (adaptive)
Gauss-Hermite quadrature.}

\item{link}{character specifying the link function.}

\item{theta}{starting values for covariance matrix.}

\item{beta}{starting values for fixed effect coefficients.}

\item{opt_func}{general optimization function to use. It
needs to have an interface like \code{\link{optim}}.}

\item{n_threads}{integer with number of threads to use.}

\item{skew_start}{starting value for the Pearson's moment coefficient of
skewness parameter when a SNVA is used. Currently,
a somewhat arbitrary value.}

\item{kappa}{numeric scalar with the penalty in the relaxed problem
ensuring the monotonicity of the survival curve.}

\item{dtformula}{\code{\link{formula}} with the derivative of the
baseline survival function.}
}
\description{
Create mgsm Object to Pass to psqn Method
}
