% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgsm-psqn.R
\name{optim_mgsm_psqn}
\alias{optim_mgsm_psqn}
\title{Optimize mgsm Object Using the psqn Method}
\usage{
optim_mgsm_psqn(
  object,
  par = NULL,
  rel_eps = sqrt(.Machine$double.eps),
  max_it = 1000L,
  n_threads = object$n_threads,
  c1 = 1e-04,
  c2 = 0.9,
  use_bfgs = TRUE,
  cg_tol = 0.2,
  strong_wolfe = TRUE,
  trace = 0L
)
}
\arguments{
\item{object}{object of class \code{"mgsm_psqn"}.}

\item{par}{initial values for the parameters to be optimized over. A
default value is used if this is \code{NULL}.}

\item{rel_eps}{relative convergence threshold.}

\item{max_it}{maximum number of iterations.}

\item{n_threads}{number of threads to use.}

\item{c1, c2, use_bfgs, cg_tol, strong_wolfe, trace}{arguments passed to
\code{\link{psqn}}.}
}
\description{
Optimize mgsm Object Using the psqn Method
}
