% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree.R
\name{optim_pedigree_psqn}
\alias{optim_pedigree_psqn}
\title{Optimize Mixed Generalized Survival Model for Pedigree Data Using the
psqn Method}
\usage{
optim_pedigree_psqn(
  object,
  par = NULL,
  rel_eps = sqrt(.Machine$double.eps),
  max_it = 1000L,
  c1 = 1e-04,
  c2 = 0.1,
  trace = 0L,
  use_bfgs = TRUE,
  cg_tol = 0.5,
  strong_wolfe = TRUE,
  max_cg = NULL,
  pre_method = 1L
)
}
\arguments{
\item{object}{object from \code{\link{make_pedigree_ADFun}}.}

\item{par}{initial values for the parameters to be optimized over. A
default value is used if this is \code{NULL}.}

\item{rel_eps}{relative convergence threshold.}

\item{max_it}{maximum number of iterations.}

\item{c1}{arguments passed to the C++ version of \code{\link{psqn}}.}

\item{c2}{arguments passed to the C++ version of \code{\link{psqn}}.}

\item{trace}{arguments passed to the C++ version of \code{\link{psqn}}.}

\item{use_bfgs}{arguments passed to the C++ version of \code{\link{psqn}}.}

\item{cg_tol}{arguments passed to the C++ version of \code{\link{psqn}}.}

\item{strong_wolfe}{arguments passed to the C++ version of \code{\link{psqn}}.}

\item{max_cg}{arguments passed to the C++ version of \code{\link{psqn}}.}

\item{pre_method}{arguments passed to the C++ version of \code{\link{psqn}}.}
}
\description{
Optimize Mixed Generalized Survival Model for Pedigree Data Using the
psqn Method
}
\seealso{
\code{\link{make_pedigree_ADFun}}
}
