% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mgsm_ADFun.R
\name{cov_to_theta}
\alias{cov_to_theta}
\alias{theta_to_cov}
\title{Maps Between a Covariance Matrix and Its Log-Cholesky Parametrization}
\usage{
cov_to_theta(cov)

theta_to_cov(theta)
}
\arguments{
\item{cov}{the covariance matrix.}

\item{theta}{numeric vector with log-cholesky parametrization.}
}
\description{
Let \eqn{C} be a \eqn{K}-dimensional covariance matrix. Let
\eqn{C = LL^\top} where \eqn{L} is the Cholesky decomposition.
Then the two functions map between \eqn{C} and
a vector containing the non-zero elements of L where the diagonal
entries have been log transformed.
}
\examples{
set.seed(1)
Sigma <- drop(rWishart(1, 4, diag(4)))
log_chol <- cov_to_theta(Sigma)
log_chol
stopifnot(all.equal(Sigma, theta_to_cov(log_chol)))

}
