% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint.R
\name{make_joint_ADFun}
\alias{make_joint_ADFun}
\title{Construct Objective Functions with Derivatives for a Joint Survival and
Marker Model}
\usage{
make_joint_ADFun(
  sformula,
  mformula,
  sdata,
  mdata,
  id_var,
  time_var,
  mknots,
  sknots,
  n_nodes = 20L,
  skew_start = .alpha_to_gamma(-1),
  opt_func = .opt_default_CG,
  n_threads = 1L,
  sparse_hess = FALSE,
  B = NULL,
  Psi = NULL,
  Sigma = NULL,
  omega = NULL,
  alpha = NULL,
  delta = NULL,
  va_par = NULL
)
}
\arguments{
\item{sformula}{\code{\link{formula}} for the time to event. The
left-hand side must be \code{\link{Surv}} object with
type "counting".}

\item{mformula}{\code{\link{formula}} for the marker process. The
left-hand side must \code{\link{cbind}}ed if there are
more than one marker.}

\item{sdata}{\code{\link{data.frame}} with the time to event data.}

\item{mdata}{\code{\link{data.frame}} with the marker data.}

\item{id_var}{name of the individual identifier. Must be present in both
\code{sdata} and \code{mdata}.}

\item{time_var}{name of the time variable in \code{mdata}.}

\item{mknots}{vector with knots for the baseline hazard or number of
knots to use.}

\item{n_nodes}{number of nodes to use with Gauss-Legendre quadrature for
the cumulative hazard and nodes to use with Gauss-Hermite
quadrature for the entropy term.}

\item{skew_start}{starting value for the Pearson's moment coefficient of
skewness parameter when a SNVA is used. Currently, a
somewhat arbitrary value.}

\item{opt_func}{general optimization function to use. It
needs to have an interface like \code{\link{optim}}.}

\item{n_threads}{integer with number of threads to use.}

\item{sparse_hess}{logical for whether to make sparse Hessian computation
available. Memory and computation time is saved if it is
\code{FALSE}.}

\item{B}{staring value for B.}

\item{Psi}{staring value for Psi.}

\item{Sigma}{staring value for Sigma.}

\item{omega}{staring value for omega.}

\item{alpha}{staring value for alpha.}

\item{delta}{staring value for delta.}

\item{va_par}{staring value for variational parameters.}
}
\description{
Construct Objective Functions with Derivatives for a Joint Survival and
Marker Model
}
